# add other resources to pipeline benchmark