## Description

Adds more resources to the pipeline benchmark test. 

Testing against an older master commit `27598ba780` using 1 and 2 CPUs.

```
LOGLEVEL="panic" go test -bench=. -run=^# -cpu=1,2 -benchtime=15s -count=5 ./sensor/tests/pipeline
```

Results in `27598ba780`:
```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
Benchmark_Pipeline     	      82	 362217531 ns/op
Benchmark_Pipeline     	      76	 374107365 ns/op
Benchmark_Pipeline     	      68	 412559199 ns/op
Benchmark_Pipeline     	      62	 403039596 ns/op
Benchmark_Pipeline     	      62	 429223487 ns/op
Benchmark_Pipeline-2   	     100	 258295991 ns/op
Benchmark_Pipeline-2   	     100	 271433119 ns/op
Benchmark_Pipeline-2   	      98	 262775850 ns/op
Benchmark_Pipeline-2   	     100	 279071768 ns/op
Benchmark_Pipeline-2   	      88	 274708586 ns/op
PASS
ok  	github.com/stackrox/rox/sensor/tests/pipeline	300.680s
```

Results in current master (with pipeline changes):
```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
Benchmark_Pipeline     	      70	 417710609 ns/op
Benchmark_Pipeline     	      63	 439635435 ns/op
Benchmark_Pipeline     	      73	 452084103 ns/op
Benchmark_Pipeline     	      61	 421160479 ns/op
Benchmark_Pipeline     	      54	 446794429 ns/op
Benchmark_Pipeline-2   	     100	 239564399 ns/op
Benchmark_Pipeline-2   	     100	 229573150 ns/op
Benchmark_Pipeline-2   	     100	 239306448 ns/op
Benchmark_Pipeline-2   	     100	 238969205 ns/op
Benchmark_Pipeline-2   	     100	 257847904 ns/op
PASS
ok  	github.com/stackrox/rox/sensor/tests/pipeline	291.256s
```

Benchstat: underperforming on a single core but performing better when two cores are available. 
```
name         old time/op  new time/op  delta
_Pipeline     396ms ± 9%   435ms ± 4%   +9.91%  (p=0.032 n=5+5)
_Pipeline-2   269ms ± 4%   241ms ± 7%  -10.48%  (p=0.008 n=5+5)
```

## Checklist
- [ ] Investigated and inspected CI test results
- [ ] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] Determined and documented upgrade steps
- [ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))

If any of these don't apply, please comment below.

## Testing Performed

- Bench tests as described above