@lvalerom 2022-11-22
> I'll run them again just in case I've messed up somewhere

Now I'm getting these results. It looks like the runs with one CPU vary a little bit.

```
name         old time/op  new time/op  delta
_Pipeline     438ms ± 2%   455ms ± 7%  +3.91%  (p=0.032 n=5+5)
_Pipeline-2   259ms ± 4%   262ms ± 8%    ~     (p=0.841 n=5+5)
```